+++ src/main/java/com/eci/ieti/model/Travel.java
@@ -1,6 +1,5 @@
 package com.eci.ieti.model;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
@@ -11,37 +10,41 @@ import java.util.List;
 public class Travel {
 
     @Id
-    @JsonIgnore
     private String id;
-
-    private long travelId;
+    private String user;
+    private List<GeneritToUserRolWeatherOrCategory> userRolList;
+    private List<GeneritToUserRolWeatherOrCategory> weatherList;
+    private List<GeneritToUserRolWeatherOrCategory> accessoriesList;
+    private List<GeneritToUserRolWeatherOrCategory> onHandList;
+    private List<GeneritToUserRolWeatherOrCategory> cleanlinessList;
+    private List<GeneritToUserRolWeatherOrCategory> shoppingList;
+    private List<GeneritToUserRolWeatherOrCategory> medicineList;
+    private List<GeneritToUserRolWeatherOrCategory> clothesList;
+    private List<GeneritToUserRolWeatherOrCategory> severalList;
+    private String destiny;
+    private String travelId;
     private String title;
     private String description;
-    private String lugar;
     private Date dueDate;
-    private String user;
-    private List<Category> category;
 
-    public long getTravelId() {
+    public Travel() {
+    }
+
+    public Travel(String id, List<GeneritToUserRolWeatherOrCategory> generitToUserRolWeatherOrCategoryList) {
+        this.id = id;
+        this.userRolList = generitToUserRolWeatherOrCategoryList;
+    }
+
+    public String getTravelId() {
         return travelId;
     }
 
-    public void setTravelId(long travelId) {
+    public void setTravelId(String travelId) {
         this.travelId = travelId;
     }
 
-
-
     public String getTitle() {
         return title;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
     }
 
     public void setTitle(String title) {
@@ -56,16 +59,20 @@ public class Travel {
         this.description = description;
     }
 
-    public String getLugar() {
-        return lugar;
-    }
-
-    public void setLugar(String lugar) {
-        this.lugar = lugar;
-    }
-
     public Date getDueDate() {
         return dueDate;
+    }
+
+    public void setDueDate(Date dueDate) {
+        this.dueDate = dueDate;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
     }
 
     public String getUser() {
@@ -76,15 +83,83 @@ public class Travel {
         this.user = user;
     }
 
-    public void setDueDate(Date dueDate) {
-        this.dueDate = dueDate;
+    public List<GeneritToUserRolWeatherOrCategory> getAccessoriesList() {
+        return accessoriesList;
     }
 
-    public List<Category> getCategory() {
-        return category;
+    public void setAccessoriesList(List<GeneritToUserRolWeatherOrCategory> accessoriesList) {
+        this.accessoriesList = accessoriesList;
     }
 
-    public void setCategory(List<Category> category) {
-        this.category = category;
+    public List<GeneritToUserRolWeatherOrCategory> getOnHandList() {
+        return onHandList;
+    }
+
+    public void setOnHandList(List<GeneritToUserRolWeatherOrCategory> onHandList) {
+        this.onHandList = onHandList;
+    }
+
+    public List<GeneritToUserRolWeatherOrCategory> getCleanlinessList() {
+        return cleanlinessList;
+    }
+
+    public void setCleanlinessList(List<GeneritToUserRolWeatherOrCategory> cleanlinessList) {
+        this.cleanlinessList = cleanlinessList;
+    }
+
+    public List<GeneritToUserRolWeatherOrCategory> getShoppingList() {
+        return shoppingList;
+    }
+
+    public void setShoppingList(List<GeneritToUserRolWeatherOrCategory> shoppingList) {
+        this.shoppingList = shoppingList;
+    }
+
+    public List<GeneritToUserRolWeatherOrCategory> getMedicineList() {
+        return medicineList;
+    }
+
+    public void setMedicineList(List<GeneritToUserRolWeatherOrCategory> medicineList) {
+        this.medicineList = medicineList;
+    }
+
+    public List<GeneritToUserRolWeatherOrCategory> getClothesList() {
+        return clothesList;
+    }
+
+    public void setClothesList(List<GeneritToUserRolWeatherOrCategory> clothesList) {
+        this.clothesList = clothesList;
+    }
+
+    public List<GeneritToUserRolWeatherOrCategory> getSeveralList() {
+        return severalList;
+    }
+
+    public void setSeveralList(List<GeneritToUserRolWeatherOrCategory> severalList) {
+        this.severalList = severalList;
+    }
+
+    public List<GeneritToUserRolWeatherOrCategory> getWeatherList() {
+        return weatherList;
+    }
+
+    public void setWeatherList(List<GeneritToUserRolWeatherOrCategory> weatherList) {
+        this.weatherList = weatherList;
+    }
+
+    public String getDestiny() {
+        return destiny;
+    }
+
+    public void setDestiny(String destiny) {
+        this.destiny = destiny;
+    }
+
+    public List<GeneritToUserRolWeatherOrCategory> getUserRolList() {
+        return userRolList;
+    }
+
+    public void setUserRolList(List<GeneritToUserRolWeatherOrCategory> generitToUserRolWeatherOrCategoryList) {
+        this.userRolList = generitToUserRolWeatherOrCategoryList;
     }
 }
