--- src/main/java/com/eci/ieti/persistence/repository/repo/UserRolRepository.java
@@ -1,9 +1,0 @@
-package com.eci.ieti.persistence.repository.repo;
-
-import com.eci.ieti.model.UserRolSelect;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-public interface UserRolRepository extends MongoRepository<UserRolSelect, String> {
-
-    public UserRolSelect findUserRolSelectById(String id);
-}
