+++ src/main/java/com/eci/ieti/services/TravelService.java
@@ -3,8 +3,8 @@ package com.eci.ieti.services;
 import com.eci.ieti.exceptions.TravelException;
 import com.eci.ieti.model.Country;
 import com.eci.ieti.model.ListCategories;
-import com.eci.ieti.model.Travel;
 import com.eci.ieti.model.GeneritToUserRolWeatherOrCategory;
+import com.eci.ieti.model.Travel;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
