+++ src/main/java/com/eci/ieti/services/impl/TravelServiceImpl.java
@@ -4,10 +4,10 @@ import com.eci.ieti.exceptions.persistence.TravelPersistenceException;
 import com.eci.ieti.model.Country;
 import com.eci.ieti.model.GeneritToUserRolWeatherOrCategory;
 import com.eci.ieti.model.ListCategories;
+import com.eci.ieti.model.Travel;
 import com.eci.ieti.persistence.repository.CustomRepository;
 import com.eci.ieti.services.TravelService;
 import com.eci.ieti.exceptions.TravelException;
-import com.eci.ieti.model.Travel;
 import com.eci.ieti.persistence.TravelPersistenceService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
